init/aws-nitro: separate nitro init mods into their own files
